requirements.txt
django
# mysqlclient
python-decouple
django-widget-tweaks
django-mptt
django-uuslug
Pillow
xmltodict